[Clean up] Move TF1 only TPU utils to r1 folder.
PiperOrigin-RevId: 279798675
